requirements.txt
lxml
requests
selenium
python-telegram-bot